// frontend/src/types/types.ts
export interface ProductResponse {
    productId: number;
    productName: string;
    productCategory: string;
    price: number;
    review: number;
    imagePath: string;
    reviewsCount: number;
    
}
// ...
export enum LoadingStatus {
    LOADED = "LOADED",
    LOADING = "LOADING",
    ERROR = "ERROR",
    NEVER = "NEVER",
    SUCCESS = "SUCCESS"
}